Fix two possible deadlocks
Both occured during slow executions of the code (activated debugger,
valgrind)

First was in the destructor of the threadpool. Deinitilizing the queue
without telling the threads to exit their main loop could lead to
threads that call tryWaitPop() again before they are destroyed. Hence
they could not join, since they are waiting for another task.

Same could happen by reducing the number of threads dynamically. But
since the queue is not deinitialized, a waiting thread could only be
woken up correctly by another submit to the queue. But if the thread
that is supposed to submit to the queue still waits for a thread to
join, this will never happen -> deadlock

There might be a better way than detaching the closed threads, but some
experiments to stop the threads from waiting in the tryWaitPop()
function failed. The detach version works fine so far and waiting threads
are killed after they can get a final task from the queue or when the
pool is destroyed and the queue calls notify_all() on the condition
variable.