Add non SSE mat4Single
This serves as comparison in Benchmarks and as backup for non SSE
systems. Interestingly the clang compiler produces similar fast
code for the single version as for the SSE version. Seems like
clang is using auto vectorization since the results are nearly as
fast as the gcc SSE results, while the gcc Single results are 4
times slower - as expected.