Optimize matXSIMD by using __m128 directly
The matXSIMD has now an array of __m128 as internal data instead
of an array of mat4.Instead of block matrix multiplication it does
the multiplications per register.
Check if there is potential for further optimizations by changing
the order of intrinsics, using temporaries or changing the register
ordering in the array to row major.