Increase number of Tasks in Threadpool test
The reason is, that the test sometimes fails on travis, because
the signal to stop a thread comes too late. When the queue is
already empty, the thread might still wait for a final task.
By adding more tasks, the probability for this to happen is smaller.
Maybe there should be a smarter way like filling the queue another
time if the thread number isn't as expected.