Add new thread pool version
This commit is a little bit messy. Target is to remove the old
thread pool as soon as the new one has all the needed functionality.
The old files serve as a lookup. Goal of the new one is to be used
together with the memory manager.
Before doing something else I should get to a point where
everything is nice and clean again.