Test for LIFO thread closing and fix a bug
Because the threads main loop was exited if mCloseThreads
was set to true, the main loop of all threads was left when
calling CloseThreads(). Only the desired number of threads
were destroyed, but the left ones have left their main loop
and therefore wasn't processing any tasks anymore. Since the
previous tests just checked the number of threads and not if
they are actually doing any work, this problem was not recognized
until now. The newly added test for LIFO thread destruction
will fail/deadlock, if this error occurs again.