Change addTest cmake macro
It is now possible to add a definition to a test by adding another
input starting with "-D". Additionally everything after a "-" in the
tests name will be removed when searching for the corresponding cpp
file. This allows reusage of the same test file with different
setups (for example controlled by definitions).