Make usage of multi-queue thread pools safer
If a thread pool has multiple queues, you have to always specify
the queue that should be used for a certain operation. Otherwise
a static assertion is triggered. Before this commit, the default
queue was the queue with index 0. This should help to avoid bugs
because of an unspecified queue number. Additionally there can't
be any false assumptions of the default behaviour -> Does
HasTasks() tell me if any queue has tasks or if the default queue
has tasks?
For single queue thread pools, everything is the same as before.