Fixed a bug in dense gauss algorithm
Matrices that have a size which is not a multiple
of the number of register values have unused entries.
These entries needs to be zero during the pivoting step,
otherwise a random value from memory could be the largest value
which results in pivot element index that is larger
than the matrix size. The loop which should prevent that was
never executed due to a wrong variable. Added a DEV_EXCEPTION
to catch invalid indices.